chore: extract shared types to package
