+++ packages/api/src/modules/smart-home/smart-lights/smart-lights.service.ts
@@ -1,20 +1,5 @@
 import { Injectable, Logger } from '@nestjs/common';
-
-export interface ISmartLight {
-  id: string;
-  name: string;
-  color: string;
-  brightness: number;
-  isOn: boolean;
-}
-
-export type ISmartLightOperation = Partial<Omit<ISmartLight, 'id' | 'name'>>;
-
-export interface ISmartLightOperationResult {
-  lightId: string;
-  success: boolean;
-  reason?: string;
-}
+import type { ISmartLight, ISmartLightOperation, ISmartLightOperationResult } from '@home-toolkit/types/smart-home';
 
 /**
  * Used to interact with a smart light setup (e.g. IKEA Trådfri or Philips Hue)
