+++ packages/types/README.md
@@ -1,0 +1,3 @@
+# @home-toolkit/types
+
+Shared types for this monorepo.
