+++ packages/types/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@home-toolkit/types",
+  "packageManager": "yarn@3.2.2",
+  "private": true
+}
