+++ packages/types/smart-home/index.ts
@@ -1,0 +1,1 @@
+export * from "./smart-lights";
