+++ packages/types/smart-home/smart-lights/index.d.ts
@@ -1,0 +1,15 @@
+export interface ISmartLight {
+  id: string;
+  name: string;
+  color: string;
+  brightness: number;
+  isOn: boolean;
+}
+
+export type ISmartLightOperation = Partial<Omit<ISmartLight, "id" | "name">>;
+
+export interface ISmartLightOperationResult {
+  lightId: string;
+  success: boolean;
+  reason?: string;
+}
