+++ packages/web/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "web",
+  "name": "@home-toolkit/web",
   "version": "0.0.1",
   "scripts": {
     "dev": "vite dev",
@@ -12,6 +12,7 @@
     "format": "prettier --write --plugin-search-dir=. ."
   },
   "devDependencies": {
+    "@home-toolkit/types": "workspace:^",
     "@sveltejs/adapter-auto": "next",
     "@sveltejs/kit": "next",
     "@typescript-eslint/eslint-plugin": "^5.27.0",
